Changed NSInteger to int so that it compiles with OS X 10.4 SDK.
git-svn-id: https://grandperspectiv.svn.sourceforge.net/svnroot/grandperspectiv/GrandPerspective/trunk@797 3e1cca9a-9203-0410-b20e-959a89a76e17

+++ code/control/EditFilterRuleWindowControl.m
@@ -147,7 +147,7 @@
 @end
 
 
-NSInteger stringCompare(id s1, id s2, void *context) {
+int stringCompare(id s1, id s2, void *context) {
   return [s1 compare: s2];
 }
